refactor: add read_rt_file.c for norm
